gojira-grab/requirements.txt
beautifulsoup4==4.12.3
jira==3.8.0
PyYAML==5.4.1
PyYAML==6.0.2
Requests==2.32.3
urllib3==1.26.5
